Eliminated unnecessary TypeConverter handling for well known types.
Added caching of compiled lambdas for the life of each query.

Before TypeConverter support was introduced:
```
Retrieved 700 entities via linq in 1599 ms.
Fetch Entities [1,601 ms]
└─ Get Entity List x100 [1,592 ms] (99.388%)
  ├─ Open Transaction x100 [543 ms] (34.100%)
  ├─ Create Repository x100 [83 ms] (5.240%)
  └─ Where & To List x100 [892 ms] (56.041%)
```

After TypeConverter support:
```
Retrieved 700 entities via linq in 12498 ms.
Fetch Entities [12,500 ms]
└─ Get Entity List x100 [12,490 ms] (99.919%)
  ├─ Open Transaction x100 [688 ms] (5.509%)
  ├─ Create Repository x100 [100 ms] (0.804%)
  └─ Where & To List x100 [11,559 ms] (92.547%)
```

After these performance improvements:
```
Retrieved 700 entities via linq in 1879 ms.
Fetch Entities [1,881 ms]
└─ Get Entity List x100 [1,872 ms] (99.495%)
  ├─ Open Transaction x100 [597 ms] (31.907%)
  ├─ Create Repository x100 [87 ms] (4.654%)
  └─ Where & To List x100 [1,096 ms] (58.546%)
```